Implemented the strict scheduler functionality
Now setting the SCHEDULER_STRICT in the Scheduler.h file, the actions should be executed about the right time (and not necessarily on an exact iteration).
This is useful when using some long cycles (say 1h), because otherwise all actions are executed with that precision.

+++ source/include/Measurement.h
@@ -5,7 +5,7 @@
 
 // structures for the meaasurement (stored in EEPROM)
 struct measurement {
-  uint16_t deltaWithLastMeasurement;
+  uint32_t deltaWithLastMeasurement;
   byte humidity[2];
   byte temperature[2];
   uint16_t voltage;
